+++ C++/dp_3d_twoByTwo.h
@@ -3,5 +3,5 @@
 #include "fcfs.h"
 #include "get_window.h"
 
-tuple<tuple<char, int, float>, tuple<char, int, float>, double> window_oneSol_dp_v1(vector<float> a, vector<float> b, vector<float> c, float W_same, float W_diff, tuple<char, int, float> last_X, tuple<char, int, float> last_Y);
-tuple<float, double> schedule_by_window_dp_v1(vector<float> a_all, vector<float> b_all, vector<float> c_all, float W_same, float W_diff, int carNum);
+tuple<tuple<char, int, double>, tuple<char, int, double>, double> window_oneSol_dp_v1(vector<double> a, vector<double> b, vector<double> c, double W_same, double W_diff, tuple<char, int, double> last_X, tuple<char, int, double> last_Y);
+tuple<double, double> schedule_by_window_dp_v1(vector<double> a_all, vector<double> b_all, vector<double> c_all, double W_same, double W_diff, int carNum);
