+++ C++/fcfs.cpp
@@ -1,13 +1,13 @@
 #include "fcfs.h"
 
-float W_same, W_diff;
+double W_same, W_diff;
 
-tuple<tuple<char, int, float>, float> schedule_single_lane(char lane, vector<float> traffic, tuple<char, int, float> prev)
+tuple<tuple<char, int, double>, double> schedule_single_lane(char lane, vector<double> traffic, tuple<char, int, double> prev)
 {
     char prevLane = get<0>(prev);
-    float prevTime = get<2>(prev);
-    tuple<char, int, float> last = make_tuple('0', 0, 0.0);
-    float wait_time = 0;
+    double prevTime = get<2>(prev);
+    tuple<char, int, double> last = make_tuple('0', 0, 0.0);
+    double wait_time = 0;
     if (prevLane == '0')
     {
         last = make_tuple(lane, 1, traffic[0]);
@@ -30,16 +30,16 @@ tuple<tuple<char, int, float>, float> schedule_single_lane(char lane, vector<flo
     return make_tuple(last, wait_time);
 }
 
-pair<float, double> first_come_first_serve_v1(float timeStep, vector<float> a_all, vector<float> b_all, vector<float> c_all)
+pair<double, double> first_come_first_serve_v1(double timeStep, vector<double> a_all, vector<double> b_all, vector<double> c_all)
 {
     auto t0 = chrono::high_resolution_clock::now();
-    vector<float> a(a_all.begin() + 1, a_all.end());
-    vector<float> b(b_all.begin() + 1, b_all.end());
-    vector<float> c(c_all.begin() + 1, c_all.end());
-    float t = 0;         // time
+    vector<double> a(a_all.begin() + 1, a_all.end());
+    vector<double> b(b_all.begin() + 1, b_all.end());
+    vector<double> c(c_all.begin() + 1, c_all.end());
+    double t = 0;         // time
     char B_prevTo = 'Y'; // Which lane the previous vehicle in lane B go to
-    float X_lastT = -W_diff;
-    float Y_lastT = -W_diff;
+    double X_lastT = -W_diff;
+    double Y_lastT = -W_diff;
     char X_lastFrom = '0';
     char Y_lastFrom = '0';
 
@@ -665,27 +665,27 @@ pair<float, double> first_come_first_serve_v1(float timeStep, vector<float> a_al
     auto t1 = chrono::high_resolution_clock::now();
     double totalComputeTime = chrono::duration_cast<chrono::nanoseconds>(t1 - t0).count();
     totalComputeTime *= 1e-9;
-    float T_last = max(X_lastT, Y_lastT);
+    double T_last = max(X_lastT, Y_lastT);
     cout << "fcfs result: " << T_last << " " << totalComputeTime << endl;
     return {T_last, totalComputeTime};
 }
 
-tuple<float, float, double> first_come_first_serve_v2(vector<float> a_all, vector<float> b_all, vector<float> c_all)
+tuple<double, double, double> first_come_first_serve_v2(vector<double> a_all, vector<double> b_all, vector<double> c_all)
 {
     auto t0 = chrono::high_resolution_clock::now();
-    vector<float> a(a_all.begin() + 1, a_all.end());
-    vector<float> b(b_all.begin() + 1, b_all.end());
-    vector<float> c(c_all.begin() + 1, c_all.end());
-    float X_lastT = -W_diff;
-    float Y_lastT = -W_diff;
+    vector<double> a(a_all.begin() + 1, a_all.end());
+    vector<double> b(b_all.begin() + 1, b_all.end());
+    vector<double> c(c_all.begin() + 1, c_all.end());
+    double X_lastT = -W_diff;
+    double Y_lastT = -W_diff;
     char X_lastFrom = '0';
     char Y_lastFrom = '0';
-    tuple<char, int, float> last_X, last_Y;
-    float first = 0;
+    tuple<char, int, double> last_X, last_Y;
+    double first = 0;
     default_random_engine generator(time(NULL));
     uniform_int_distribution<int> distribution(0, 1);
-    float total_wait = 0;
-    float wait_time = 0;
+    double total_wait = 0;
+    double wait_time = 0;
     int vehicle_num = a_all.size() + b_all.size() + c_all.size() - 3;
 
     while (a.size() > 0 || b.size() > 0 || c.size() > 0)
@@ -859,8 +859,8 @@ tuple<float, float, double> first_come_first_serve_v2(vector<float> a_all, vecto
         // cout << "last_X: " << get<0>(last_X) << " " << get<1>(last_X) << " " << get<2>(last_X) << endl;
         // cout << "last_Y: " << get<0>(last_Y) << " " << get<1>(last_Y) << " " << get<2>(last_Y) << endl;
     }
-    float T_last = max(max(get<2>(last_X), get<2>(last_Y)), max(X_lastT, Y_lastT));
-    float T_delay = total_wait / vehicle_num;
+    double T_last = max(max(get<2>(last_X), get<2>(last_Y)), max(X_lastT, Y_lastT));
+    double T_delay = total_wait / vehicle_num;
     auto t1 = chrono::high_resolution_clock::now();
     double totalComputeTime = chrono::duration_cast<chrono::nanoseconds>(t1 - t0).count();
     totalComputeTime *= 1e-9;
