+++ C++/group_milp.h
@@ -6,8 +6,8 @@
 #include "get_group.h"
 using namespace std;
 
-extern float W_same, W_diff;
-vector<pair<int, int>> fixed_threshold_grouping(vector<float> &traffic, float timeStep);
-tuple<float, float, double> solve_group_milp(vector<float> A, vector<float> B, vector<float> C, float timeStep);
+extern double W_same, W_diff;
+vector<pair<int, int>> fixed_threshold_grouping(vector<double> &traffic, double timeStep);
+tuple<double, double, double> solve_group_milp(vector<double> A, vector<double> B, vector<double> C, double timeStep);
 
 #endif
